Fixed: Hoisting together with source map generation crashed as StringBuffer does not have the attribute 'atoms'
When a 'var' declaration later in the code hides a instance variable the compiler removes the string 'self.'
in front of all the identifiers as it is now a local variable. This changes moves the removal to the class
StringBuffer so it can be handles differently when generating source maps or not.